Move the expanding row component to the shared components folder.
