+++ jujugui/static/gui/src/app/assets/css/_components.scss
@@ -62,7 +62,7 @@
 @import '../../components/entity-details/entity-details';
 @import '../../components/entity-details/header/header';
 @import '../../components/expanding-progress/expanding-progress';
-@import '../../components/expanding-row/expanding-row';
+@import '../../components/shared/expanding-row/expanding-row';
 @import '../../components/expert-block/expert-block';
 @import '../../components/expert-card/expert-card';
 @import '../../components/expert-contact-card/expert-contact-card';
