+++ jujugui/static/gui/src/app/components/budget-table/row/test-row.js
@@ -5,7 +5,7 @@ const React = require('react');
 const enzyme = require('enzyme');
 
 const BudgetTableRow = require('./row');
-const ExpandingRow = require('../../expanding-row/expanding-row');
+const ExpandingRow = require('../../shared/expanding-row/expanding-row');
 const GenericButton = require('../../generic-button/generic-button');
 
 describe('BudgetTableRow', function() {
