+++ jujugui/static/gui/src/app/components/payment/details/address/address.js
@@ -5,7 +5,7 @@ const PropTypes = require('prop-types');
 const React = require('react');
 
 const AddressForm = require('../../../address-form/address-form');
-const ExpandingRow = require('../../../expanding-row/expanding-row');
+const ExpandingRow = require('../../../shared/expanding-row/expanding-row');
 const GenericButton = require('../../../generic-button/generic-button');
 const initUtils = require('../../../../init/utils');
