+++ jujugui/static/gui/src/app/components/profile/credential-list/credential-list.js
@@ -8,7 +8,7 @@ const shapeup = require('shapeup');
 const BasicTable = require('../../basic-table/basic-table');
 const ButtonDropdown = require('../../button-dropdown/button-dropdown');
 const CredentialAddEdit = require('../../credential-add-edit/credential-add-edit');
-const ExpandingRow = require('../../expanding-row/expanding-row');
+const ExpandingRow = require('../../shared/expanding-row/expanding-row');
 const GenericButton = require('../../generic-button/generic-button');
 const ProfileCredentialListDelete = require('./delete/delete');
 const Spinner = require('../../spinner/spinner');
