+++ jujugui/static/gui/src/app/components/shared/expanding-row/_expanding-row.scss
