+++ jujugui/static/gui/src/app/components/shared/expanding-row/expanding-row.js
