+++ jujugui/static/gui/src/app/components/shared/expanding-row/test-expanding-row.js
